[gnsdk] Don't create c++ bindings for zx fidl
I overlooked this in the initial migration and it caused a roller
error. This brings the gn sdk into alignment with the in-tree sdk.
https://cs.opensource.google/fuchsia/fuchsia/+/main:build/cpp/fidl_cpp.gni;l=301

Not sure why this wasn't caught in any of the lsc builds, but I'll
follow up this CL with an attempt to add this case to the in-tree
gn sdk tests.

Bug: 115033
Change-Id: Ice5906484819fd2b8844264d7e924bb0804007e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796990
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>