[ffx][schema] Remove incomplete schema command
We'll want to replace this with a version that's built on the new
subtool macros and uses a typesafe source of info for the type of the
structured output.

This also sets us up for not regressing more than the machine output
mechanism already has by enforcing type-safety on the argument to the
machine type, in some cases removing invalid machine declarations where
no attempt was made to actually implement machine output.

There will be some follow-up work to make things more correct in
general.

Bug: 121214
Fixed: 121213

Change-Id: I56ea740c5e0fd20d2b8bcd04ec4df0ce2bd62557
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796088
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>