[async_patterns] Receiver and Callback
The primary motivation is to support calling methods that return a
value on some object managed by a DispatcherBound.

When a parent object manages a child object via DispatcherBound, these
two objects are intended to live in different concurrency domains.
Meaning the child could back a response to the parent as the parent
is entering its destructor. In order to make this safe, we make the
child send responeses via task posting, and passivate the tasks if the
parent is destroyed. At the same time, we require the parent to live on
a synchronized dispatcher. This way the destruction of the parent will
be synchronized with those posted response tasks. All these is
encapsulated in an `async_patterns::Receiver` object, documented in the
receiver.h.

The restrictions on sending arguments is carved out into a dedicated
header, and used by both Receiver and DispatcherBound.

A unit test is added to DispatcherBound showing how to call methods
asynchronously with non-void returns.

Multiply: fuchsia-pkg://fuchsia.com/async_patterns_tests#meta/async_patterns_tests.cm
Bug: 119641
Change-Id: I80f6f2a694763835399449fd6a6f8d243ae9664a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792456
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>