[starnix] Change mounts to component_mounts
This updates the configuration values for mounts that are parsed on
behalf of a component running in a galaxy. This prevents collisions
between the galaxy and component configurations, which is especially
useful in tests, where the .cml of the test contains both galaxy
configuration and component configuration values.

Change-Id: If475b7c0fa366b5d40061495f05dfd8389b842f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797982
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>