[zxio] Maintain valid LRU on cache update
This change fixes two aspects of the LRU cache eviction behavior in the
RouteCache such that it always:
 - removes out-of-date entries from the LRU list
 - only moves validated entries to the front of the list

Currently, if an entry in the RouteCache is invalidated and then a
client attempts to retrieve it, the refreshed entry will overwrite the
invalid one in the cache, but the invalid entry will not be removed from
the LRU list; effectively a duplicate will be added leaving a dangling
node. This can eventually lead to a panic when that node becomes the
least-recently-used and is removed, and the cache tries to also remove
the (nonexistent) corresponding cache entry.

To fix this, always remove a refreshed entry from the LRU list before
overwriting it in the cache.

Also, only mark an entry as the most recently used after its
eventpair(s) have been validated.

Fixed: 119098
Test: zxio-unittest-package
Change-Id: I9bf05234ec1e49f90933f2b85424c8a58c5f9290
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796130
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>