[fake-clock] Ignored deadlines never expire
Add a new API endpoint for the fake-clock to make deadlines never
expire. If a user of fake-clock calls ignore_named_deadline prior
to adding a named deadline, the deadline will have its expiration
to infinite.

Includes tests to ensure that there is a no-op when trying to
ignore the same deadline twice and that the deadline is properly
overridden.

Change-Id: If064f55f9c96618807dfc3afd7f90734bfa674f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792731
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>