[overnet] Update circuit routes when new connections occur
In order for a circuit node to connect to another node, it needs to
receive a routing update identifying that node and how it can be
reached. In order for a circuit node to *be connected to* by another
node, it doesn't need anything; a connection comes in, identifies the
node it came from, and all is well. This can occur if a new node's
connection request to another node manages to race ahead of the routing
updates produced by that node joining the network. As a result, a node
can receive a connection from another node while not being able to
connect back to that node. This is bad for Overnet, which needs to be
able to freely establish connections to move handles around.

The solution is to add a route to the routing table when we get a
connection from a node we don't recognize. We have to score it low on
link quality because we don't know how fast it is without the quality
information that comes in a routing update, but as long as it's present
we can connect back to any node that connects to us.

Fixed: 120884
Bug: 119353
Change-Id: Iadd3631a244162a8f9f0cb5bc21153c31b13af08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797149
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>