[ffx] Add tracing::instrument attributes
Adds spans to functions that use tracing.

The primary advantage is that log message will provide the name of the function. A secondary advantage is that, when we switch to a more sophisticated log-viewer such as Jaeger,we will automatically benefit from the span structure

(Also removes "level = info" where it was specified, since that is the default.)

Fixed: 119970
Change-Id: I0c66cf4879c475ba846c52aa86c6d45bef48f26c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797943
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>