[ffx][debug] List symbol index using JSON.
This makes the output of ffx debug symbol-index list machine-readable,
which allows developer environment setup logic to convert it into flags
and arguments for CLI utilities that can use it.

For example, the work in fxb/111998 would use this to produce
--debug-file-directory flags for llvm utilities via jq in fx wrappers,
and it could similarly be used to configure build ID directories for
GDB and/or LLDB for kernel use.

Bug: 111998
Change-Id: Ia90915005d3b5fac1580a9e5d753c81b6216102a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798382
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Daniel Thornburgh <dthorn@google.com>