[fixit][storage] remove some smklein TODOs
Remove some TODOs attributed to smklein which we are not going to do.

They are removed because they are stale, or test coverage is fine, or
because things have been working fine without the fix, or they are
related to minfs so we are never going to fix.

Change-Id: Ia64dc499d3e9f95f41c5d76ad910d2f55c9a902f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799062
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>