[libzircon_boot] add support for slotless boot
Adds support for a new "slotless" boot mode, another option in addition
to the existing A/B/R or force-recovery boot modes.

This is primarily intended for upcoming support for RAM-booting, which
will be slotless since it doesn't use A/B/R metadata at all, but may
also be useful in early bring-up for boards that don't have A/B/R
support yet.

Bug: b/267627997
Change-Id: Icb796e8492fbd15bfda66512cc5eb5cdcef95972
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797787
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>