[ffx] Delete ValueStrategy and ConfigValue type conversion for Map type
For a ConfigQuery, to get a Map<String, Value> typed config value, it
will call `validate_type::<Map<String, Value>>` on the map recursively,
which tries converting every single value (even if it's not a Object)
in the map to Map<String, Value>, and it will fail.

So every ConfigQuery::get<Map<String, Value>> call will only return
an Error instead of the desired value. The only correct way to get a
Map<String, Value>> is to get a Value and then check if it's a
Value::Object.

The code is currently not used anywhere so it's safe to delete it.

Change-Id: I2ef8784ab0e9ca6a12f69ee4b02e9a7e92454285
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799206
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>