[dhcp] Don't use std::mem::size_of_val
std::mem::size_of_val is a footgun, because it is easy to accidentally
use it to measure the size of a Vec (and get back 24 regardless of the
Vec's length), when you meant to measure the size of the slice contained
by the Vec (which is the number of bytes taken up by the data stored in
the slice).

This was causing us to incorrectly serialize the ParameterRequestList
and PathMTUPlateauTable DHCP options.

Test: dhcp_protocol
Change-Id: I22ae5825d41849e4a8a5a67fcd731ecece9a817e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799684
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>