[starnix] Fix binder driver refcount issue.
The binder driver allowed for weak and strong refcount operation to be
re-ordered. It means that the following operations:
- IncStrong IncWeak decStrong decWeak
could be reordered into:
- IncStrong decStrong IncWeak decWeak
This is invalid, as the remote object would be freed after the decStrong
command.

To ensure this doesn't happen, the binder driver will not send any
decrement command until all increment commands (weak and strong) are
acknowledged. This ensires in the previous sequence that the decStrong
command cannot be executed beffore the IncWeak happened.

With this, the binderLibTest is not crashing anymore on an use after
free.

Fixed: 120058
Change-Id: I06bc03ff8933c9597db350dea54550bca985c16d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800482
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>