[base-resolver] Swap in the true-base-resolver
Replaces the current base-resolver implementation with that of the
true-base-resolver (which serves package directories directly from
blobfs instead of relying on pkg-cache's pkgfs/packages directory).

Uses this new base-resolver to resolve pkg-cache and so deletes
pkg-cache-resolver.

Serves the shell-commands "bin" directory from base_resolver instead
of pkg-cache. console-launcher doesn't add the "bin" directory to the
shell's namespace until the "bin" directory starts responding to
requests (to prevent hanging the shell with unresponsive namespace
entries). base_resolver using pkgfs/packages caused pkg-cache to start
early in the boot sequence, but with this change there are some
configurations where the shell accessing the "bin" directory is the
first use of a pkg-cache capability and so is when CF starts
pkg-cache. On asan builds this is slow enough that some tests fail
because "bin" isn't added to the shell's namespace soon enough and so
some shell binaries cannot be found.

Fixed: 101492
Change-Id: I188d089fe5cabb9874807018ad8b753172679100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794970
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Sen Jiang <senj@google.com>