[zircon][docs][fixit]Fixing markdown syntax in boot-options.md
This is a reland of fxr/799643: Revert "[docs][fixit]Escaping brackets which are not ref links"

This escapes the [ so that they are not interpreted as auto reference
links.

I also added comments to the generated markdown to point to where
the file is generated.

Bug: 106059
Change-Id: I385b6d684d22603bf0625634a030059b22023879
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799822
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>