[bazel_sdk] Replace debug-symbol related bash scripts with python ones.
This replaces elf_strip.sh and generate_symbols_dir.sh with
equivalent Python script. This is more portable and removes the
last use of the `file` command which is problematic on CQ bots
and remote builds.

This also introduces an `elf_info` Python library to query ELF files
(only ELF64 is supported for now), used by both scripts.

This does not seem to slow down local or remote builds, but fixes
remote builds which failed because of the missing `file` command.

```

$ fx bazel clean && \
  fx bazel build //build/bazel/examples/hello_cpp:pkg.debug_symbols

  BEFORE (Best of 3 runs)

    INFO: Elapsed time: 17.964s, Critical Path: 11.16s
    INFO: 457 processes: 305 internal, 152 linux-sandbox.

  AFTER: (Best of 3 runs)

    INFO: Elapsed time: 17.587s, Critical Path: 12.93s
    INFO: 457 processes: 305 internal, 152 linux-sandbox.

$ fx bazel clean && \
  fx bazel build --config=remote //build/bazel/examples/hello_cpp:pkg.debug_symbols

  BEFORE (Best of 3 runs after warmup one)

    INFO: Elapsed time: 11.237s, Critical Path: 9.52s
    INFO: 457 processes: 152 remote cache hit, 305 internal.

  AFTER (Best of 3 runs after warmup one)

    INFO: Elapsed time: 11.626s, Critical Path: 9.99s
    INFO: 457 processes: 152 remote cache hit, 305 internal.
```

Fixed: 99458
Change-Id: I80bda29170d4a79d57d8795457b0d9e6f428d117
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795084
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>