[fxfs] Enable LTO
Re-land of fxrev.dev/791112 with a slightly different approach, which is
used by ffx and should work.

This gains back the performance which was lost when splitting Fxfs into
several crates.  See, for example, WalkDirectoryTreeWarm.

The cost is a ~45% slowdown in compilation time for fxfs (19s -> 33s on
my machine).  If this ends up being problematic for productivity, we
could further split up Fxfs into a binary variant used for tests (which
won't have LTO) and another used for benchmarks and production (which
would have LTO), but that doesn't seem necessary given the build time is
still reasonable.

Note that LTO is not applied to the unit-test suite, so building the
unit-test suite won't be made slower.

Bug: 119533
Change-Id: I90c7e7d10b06b38d71c32125884838586b1b1ca6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799622
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>