[tefmocheck] Only check ffx timeout error on failed tasks.
If ffx fails to connect to the target, we retry and if it succeeds, then
we keep testing. Tefmocheck should ignore the timeout log in these
cases. If ffx continues to fail to connect to the target, it should
return a fatal error and fail the task, in which case tefmocheck can
assign it this failure mode.

Change-Id: I45152193996b292cffa546545229255be4c7c6dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802008
Reviewed-by: Andrés Villegas <andresvi@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>