[ffx][writer] New writer library
This implements a new type-safe writer library for FHO-based subtools,
that more or less resembles the old interface but enforces a single type
as its output.

This will allow us to produce much more useful schema information when
we get to it, and ensure that in machine mode tools are only outputting
the type of data they claim to support.

Bug: 114369

Change-Id: I9cc9bb5554c62d58881b77c6e05e532d4ad24b6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797523
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>