[libc] Add macro for disabling SCS and hwasan instrumentation
The core-c11-thread test would fail with
https://reviews.llvm.org/D139377 because that patch would cause the
__hwasan_load/store functions to be non-leaf functions, meaning they are
instrumented with SCS. This particular test would fail with a page fault
because we could end up in the state where a thread is explicitly
detached after it has exited, meaning any hwasan checks done in the
detach call will call into hwasan and access the SCS even thought it was
deallocated when the thread exited.

This isn't indicative of anything funadmentally wrong with hwasan, but
more happenstance that we use runtime calls for access checking with
only hwasan. It's also only in the particular unique situation when we
can call into an SCS instrumented function after SCS was deallocated.

This CL adds a macro for explicitly disabling SCS and hwasan on specific
libc functions where this case is seen. It also replaces all libc uses
of __NO_SAFESTACK.

Bug: 115924, 57112
Change-Id: I959cb0df5d8bef7e7547fb73d6dcedfe5549d923
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766781
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>