[sysmem2] change display_size to display_rect
Since some codecs support cropping on the top left, it makes
sense to use a rect with upper-left values as well, instead of
a size with only the width and height.

Nothing about this CL should be taken to imply that we intend
to support top-left cropping in any pipeline that doesn't
actually use it. While non-zero top-left is possible, it's
definitely not common. Ignoring these fields when they're not
actually zero would look wrong, but wouldn't be fatal, and
would be no worse than not having the x, y top-left fields.

Change-Id: I3dd9ae74bff5538e8a7b11fdb9c5eba2157af0ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796763
Reviewed-by: John Bauman <jbauman@google.com>
API-Review: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>