[i2c][amlogic] Fix read size
Correct the read size so that only up to 8 bytes are written to the user
buffer depending on the number of read tokens issued. This corrects a
mistake in a previous CL, but should have no functional impact other
than increasing read performance slightly.

Test: fx test aml-i2c-test
Change-Id: I21f30c02d30a4294af732d2c56ed6847c77c301e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802487
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>