[fxfs] Fix lock race on rename.
There was a race where if there was no target file at first check it
would remove the object from the directory without having the correct
lock on the target that arrived there after. Now the acquire lock method
now will always return the transaction, but indicates if the object is
there or not while holding a lock on the directory to prevent that from
changing.

Bug: 119138
Change-Id: If2e1245595373be9fc6848b2313ab5708d11b525
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802428
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Tam <godtamit@google.com>