[netsvc] Avoid opening /dev with rights
Dealing in channels has the effect of eagerly encountering errors, which
required some restructuring. In particular, the previous code opened
"/dev" as a file descriptor which works even when the namespace contains
"/dev/class/network" such that "/dev" is a local node in fdio; this
would later fail in `Paver::OpenDataSink` which tries to
`component::ConnectAt` through this file descriptor which might not be
backed by a channel - it must have been the case that this code path is
not exercised when netsvc is running in core (as opposed to bringup).

Bug: 101092
Change-Id: I287f40925ee13928da7f454d113d12016cc162dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802072
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>