[ffx][overnet]: Cleanup ssh connections
When dropping an Overnet host pipe connection, explicitly kill the
spawned ssh connection. Previously this was not killed when the target
was dropped from the colleciton, causing zombie ssh connections to the
target to persist. This would require a daemon kill to remove (or a kill
on the ssh processes themselves)

Change-Id: I51dc87c3c1b27165762edf3612a42f385455ab5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802188
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Steven Grady <slgrady@google.com>