[display] Close BufferCollectionToken if import is not supported.
This change changes the default behavior of banjo ImportBufferCollection
method -- if it is not supported, the driver should try to close the
BufferCollectionToken. Otherwise it will trigger a ZX_ERR_PEER_CLOSED
in sysmem and cause a failure on all the other shared tokens.

Bug: 121411
Change-Id: I57d1fbed0d7f0359379a6216765635b95c8db3f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803770
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>