[fxfs] Add transaction lock requirement for objects.
In order to add a mutation to replace_or_insert() any object the
transaction must have an exclusive lock on that object. Exceptions are
made for objects that were newly inserted in the transaction, as those
will be exclusively owned by the transaction anyways.

Bug: 119138

Change-Id: Ic4aab42e9f8520c5c8899fd3bcaa2546c36004e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788621
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>