[ramdevice_client] Replace get_path with as_dir
Rather than exposing the block_path of the ramdisk, we now force
callers to interact with the ramdisk via either a DirectoryProxy or a
ControllerProxy, only providing the DirectoryProxy when necessary.
There is further work to be done here with regards to multiplexing
between the Controller and Block protocol.

Change-Id: I44ec41b46ea9cd723d55970368c72d6270955621
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794202
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>