[starnix] Fix ffx starnix adb command
Previously the kernel was not serving the galaxy controller in the
outgoing directory of the galaxy. Instead, it was only serving it in the
outgoing directory of the kernel itself. This used to work, when the
galaxy and starnix kernel were the same component.

Change-Id: I81ec1132dddfececea6f21418323a919917eefac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804127
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>