[netstack3] Split capacity and target_capacity
Use a separate method for reporting the target capacity of a socket's
send buffer. This fixes a bug where, if a TCP socket was full, a
reduction in its send buffer size wasn't reflected by the next call to
read its send buffer size. Add an integration regression test for the
behavior.

Since the test makes additional use of the ability to set and retrieve
the send buffer size for a TCP socket, add methods for such to the
fuchsia_async crate.

Fixed: 121449
Change-Id: I0f41af1761e1c5cd7534a1d89d690f53035a1587
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802347
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>