[session-manager] Add a core shard for `minimal`.
A variety of protocols listed in 'required_capabilites` aren't
necessarily available on products based on minimal.gni. So I've moved
them to "optional_capabilities" and tagged them with
`source_availability: "unknown"`.

I also renamed `required_capabilities` to `all_products_capabilities`
and `optional_capabilities` to `some_products_capabilities`. The use
of the word "optional" in the name of that file contributed to my
incorrect assumption that the capabilities listed in that file were
received as "optional" by the session.  This is maybe more verbose,
but I think truer to the intent of the files.

Bug: 121597

Change-Id: I05c835b2c104616d385f2b9635e0c5628373edad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800704
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>