[botanist][ffxutil] Set config values before starting the daemon.
The daemon might not be picking up new config settings if they are set
after the daemon is started, so we should try to set it before where
possible.

I also added a function to get a copy of an FFXInstance with a different
associated target so that we don't need to duplicate the number of
`config set` calls for each target.

Change-Id: I86d5c32eacd8fcb6ad150d11178245f1307e74df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804331
Reviewed-by: Andrés Villegas <andresvi@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>